Move some shared functionality into a trait
